refactor: :recycle: отделить context и локальные факты
Расширения контекста теперь принадлежат подсистемам, которые их используют, а локальные факты зависят только от узких листовых контрактов. Это убирает шесть модулей из циклических компонент без изменения поведения.
