+++ packages/core/metadata/orchestration/property/importYamlTypes.ts
@@ -1,17 +1,17 @@
 import type { ConfigurationContext, ConfigurationContextFromXML, ExternalFileEntry } from "../../context/types"
-import type { MetadataTargetOwner } from "../metadataTarget/types"
 import type { FormDataPathIndex } from "../dataPath/formIndex"
-import type { YamlDiagnosticLocation, YamlPath } from "../../validation/yamlLocations"
-import type { LocalIndexes, LocalIndexesCollector, LocalMetadataIndex } from "../../project/localIndexes"
+import type { YamlPath } from "../../validation/yamlLocations"
+import type {
+  DeferredRulePathSegment,
+  LocalIndexes,
+  LocalIndexesCollector,
+  LocalMetadataFactsWriter,
+  LocalYamlFact,
+} from "./localFacts"
 import type { MetadataItemRule, PropertyRule } from "./types"
 import type { DeferredValuePath } from "./deferredObjectValues"
 
 export type { DeferredValuePath } from "./deferredObjectValues"
-
-export interface DeferredRulePathSegment {
-  propertyKey: string
-  nestedItemType?: string
-}
 
 export interface DirectImportTraversal {
   yamlPath: YamlPath
@@ -75,26 +75,19 @@ export interface DirectImportResult {
   generatedFiles: ExternalFileEntry[]
 }
 
-export interface LocalYamlFact {
-  yamlPath: YamlPath
-  rulePath: readonly DeferredRulePathSegment[]
-  rule: PropertyRule
-  value: unknown
-  source?: YamlDiagnosticLocation
-  metadataTargetOwner?: MetadataTargetOwner
-}
-
-export interface LocalMetadataFactsWriter {
-  setOwnerFact(role: string, value: unknown): void
-  setMetadataTargetValues(values: readonly { value: string; yamlPath: YamlPath }[]): void
-}
-
 export type CollectLocalFactsFromYAMLFunction = (params: {
   fact: LocalYamlFact
   writer: LocalMetadataFactsWriter
 }) => void
 
-export { type LocalIndexes, type LocalIndexesCollector, type LocalMetadataIndex }
+export type {
+  DeferredRulePathSegment,
+  LocalIndexes,
+  LocalIndexesCollector,
+  LocalMetadataFactsWriter,
+  LocalMetadataIndex,
+  LocalYamlFact,
+} from "./localFacts"
 
 export type ImportFromXMLToYAMLFunction = (params: {
   context: ConfigurationContextFromXML
