+++ packages/core/metadata/orchestration/property/ownerFactRole.ts
@@ -1,0 +1,18 @@
+export type OwnerFactRole =
+  | "type"
+  | "attributes"
+  | "dimensions"
+  | "resources"
+  | "addressingAttributes"
+  | "tabularSections"
+  | "standardAttributes"
+  | "owners"
+  | "task"
+  | "registerRecords"
+  | "chartOfAccounts"
+  | "extDimensionTypes"
+  | "accountingFlags"
+  | "extDimensionAccountingFlags"
+  | "commonAttributeOwnerLinks"
+  | "registerType"
+  | "commands"
